Small bugfixing regarding unecessary event array

+++ Backend/Routes/vector_routes.py
@@ -40,7 +40,7 @@ def get_matching_vector_by_category(payload):
     matching_users = []
     for user in user_matches:
         user_profile = find_profile_by_id(user["id"])
-        score = round(int(user["distance"]),2) * 100
+        score = round(float(user["distance"]),2) * 100
         user_profile["match_score"] = score
         matching_users.append(user_profile)
     return matching_users, 200
